requirements.txt
python-dotenv
databricks-sql-connector
pandas
numpy
streamlit
plotly
certifi
openai
apscheduler
kaleido
requests
slack_bolt
slack_sdk
